+++ packages/apps/dev-wallet/src/App/routes.tsx
@@ -56,12 +56,7 @@ const Redirect: FC<
 export const Routes: FC = () => {
   const { isUnlocked } = useWallet();
   const isLocked = !isUnlocked;
-  const [origin, setOrigin1] = useState('/');
-
-  const setOrigin = (value: string) => {
-    console.log('setOrigin', value);
-    setOrigin1(value);
-  };
+  const [origin, setOrigin] = useState('/');
 
   const routes = createRoutesFromElements(
     <Route
@@ -77,7 +72,10 @@ export const Routes: FC = () => {
             path="/unlock-profile/:profileId"
             element={<UnlockProfile />}
           />
-          <Route path="/import-wallet" element={<ImportWallet />} />
+          <Route
+            path="/import-wallet"
+            element={<ImportWallet setOrigin={setOrigin} />}
+          />
         </Route>
       </Route>
       <Route