+++ packages/apps/dev-wallet/src/pages/transaction/Transaction.tsx
@@ -64,14 +64,15 @@ export function Transaction() {
           ...tx,
         } as ITransaction;
         await transactionRepository.updateTransaction(updated);
-        const updatedList = Txs.map((t) =>
-          t.uuid === updated.uuid ? updated : t,
-        );
-        const [overallStep, firstTx] = getOverallStep(updatedList);
-        setTxs(updatedList);
-        setSelectedTxIndex(firstTx);
-        setStep(overallStep);
-        return updatedList;
+        return loadTxs(groupId!);
+        // const updatedList = Txs.map((t) =>
+        //   t.uuid === updated.uuid ? updated : t,
+        // );
+        // const [overallStep, firstTx] = getOverallStep(updatedList);
+        // setTxs(updatedList);
+        // setSelectedTxIndex(firstTx);
+        // setStep(overallStep);
+        // return updatedList;
       }
     },
     [transaction, Txs],
@@ -81,19 +82,28 @@ export function Transaction() {
     if (!groupId) return;
     const client = createClient();
     if (!list.every(isSignedCommand)) return;
-    await Promise.all(
+    const preflightValidation = await Promise.all(
       list.map((tx) =>
         client
           .preflight({ cmd: tx.cmd, sigs: tx.sigs, hash: tx.hash } as ICommand)
-          .then((result) => {
-            if (result.result.status !== 'success') {
-              throw (result.result.error as any).message || result.result.error;
-            }
+          .then(async (result) => {
+            await transactionRepository.updateTransaction({
+              ...tx,
+              status: 'preflight',
+              preflight: result,
+              request: undefined,
+            });
+            return result.result.status === 'success';
           }),
       ),
     );
+    if (preflightValidation.some((isValid) => !isValid)) {
+      setError('Preflight failed');
+      return;
+    }
+    const listForSubmission = await loadTxs(groupId);
     await Promise.all(
-      list.map((tx) =>
+      listForSubmission.map((tx) =>
         client
           .submitOne({ cmd: tx.cmd, sigs: tx.sigs, hash: tx.hash } as ICommand)
           .then(async (request) => {